refactor

+++ src/main/java/it/unisalento/music_virus_project/billing_service/domain/entity/fee/Tax.java
@@ -3,9 +3,11 @@ package it.unisalento.music_virus_project.billing_service.domain.entity.fee;
 import it.unisalento.music_virus_project.billing_service.domain.enums.FeeType;
 import it.unisalento.music_virus_project.billing_service.domain.enums.TaxEnum;
 
+import java.math.BigDecimal;
+
 public class Tax extends FeePlan {
     private TaxEnum taxName;
-    private Double percentageOnTotal;
+    private BigDecimal percentageOnTotal;
 
     public Tax() {
         super();
@@ -18,10 +20,10 @@ public class Tax extends FeePlan {
     public void setTaxName(TaxEnum taxName) {
         this.taxName = taxName;
     }
-    public Double getPercentageOnTotal() {
+    public BigDecimal getPercentageOnTotal() {
         return percentageOnTotal;
     }
-    public void setPercentageOnTotal(Double percentageOnTotal) {
+    public void setPercentageOnTotal(BigDecimal percentageOnTotal) {
         this.percentageOnTotal = percentageOnTotal;
     }
 }
